refactor(runloop) move functionality from init to handler
This change moves a lot of functions and variables from kong/init.lua to
kong/runloop/handler.lua

This is in preparation for the async reloading of plugins and routes

Most functions have been cut-and-pasted from one place to the other,
although there are slight differences:

* A new function called runloop.set_init_versions_in_cache was added,
with functionality moved there from init.lua
* All plugin-related functions, and aux functions, have been moved to
handler.lua
