+++ kong/runloop/handler.lua
@@ -19,6 +19,7 @@ local singletons   = require "kong.singletons"
 local certificate  = require "kong.runloop.certificate"
 local concurrency  = require "kong.concurrency"
 local ngx_re       = require "ngx.re"
+local BasePlugin   = require "kong.plugins.base_plugin"
 
 
 local kong         = kong
@@ -60,8 +61,10 @@ local ERROR        = ngx.ERROR
 local CACHE_ROUTER_OPTS = { ttl = 0 }
 local SUBSYSTEMS = constants.PROTOCOLS_WITH_SUBSYSTEM
 local EMPTY_T = {}
+local TTL_ZERO = { ttl = 0 }
 
 
+local get_plugins, build_plugins, update_plugins
 
 local get_router, build_router
 local server_header = meta._SERVER_TOKENS
@@ -322,6 +325,158 @@ local function register_events()
       log(ERR, "failed broadcasting upstream ", operation, " to workers: ", err)
     end
   end)
+end
+
+
+do
+  local plugins
+  local loaded_plugins
+
+
+  local function get_loaded_plugins()
+    local loaded = assert(kong.db.plugins:get_handlers())
+
+    if kong.configuration.anonymous_reports then
+      reports.configure_ping(kong.configuration)
+      reports.add_ping_value("database_version", kong.db.infos.db_ver)
+      reports.toggle(true)
+
+      loaded[#loaded + 1] = {
+        name = "reports",
+        handler = reports,
+      }
+    end
+
+    return loaded
+  end
+
+
+  local function should_process_plugin(plugin)
+    local c = constants.PROTOCOLS_WITH_SUBSYSTEM
+    for _, protocol in ipairs(plugin.protocols) do
+      if c[protocol] == subsystem then
+        return true
+      end
+    end
+  end
+
+
+  build_plugins = function(version)
+    loaded_plugins = loaded_plugins or get_loaded_plugins()
+
+    local new_plugins = {
+      map = {},
+      combos = {},
+      loaded = loaded_plugins,
+    }
+
+    if subsystem == "stream" then
+      new_plugins.phases = {
+        init_worker = {},
+        preread     = {},
+        log         = {},
+      }
+
+    else
+      new_plugins.phases = {
+        init_worker   = {},
+        certificate   = {},
+        rewrite       = {},
+        access        = {},
+        header_filter = {},
+        body_filter   = {},
+        log           = {},
+      }
+    end
+
+    for plugin, err in kong.db.plugins:each(1000) do
+      if err then
+        return nil, err
+      end
+
+      if should_process_plugin(plugin) then
+        new_plugins.map[plugin.name] = true
+
+        local combo_key = (plugin.route    and 1 or 0)
+                        + (plugin.service  and 2 or 0)
+                        + (plugin.consumer and 4 or 0)
+
+        new_plugins.combos[plugin.name] = new_plugins.combos[plugin.name] or {}
+        new_plugins.combos[plugin.name][combo_key] = true
+      end
+    end
+
+    for _, plugin in ipairs(loaded_plugins) do
+      if new_plugins.combos[plugin.name] then
+        for phase_name, phase in pairs(new_plugins.phases) do
+          if plugin.handler[phase_name] ~= BasePlugin[phase_name] then
+            phase[plugin.name] = true
+          end
+        end
+
+      else
+        if plugin.handler.init_worker ~= BasePlugin.init_worker then
+          new_plugins.phases.init_worker[plugin.name] = true
+        end
+      end
+    end
+
+    plugins = new_plugins
+
+    return true
+  end
+
+
+  update_plugins = function()
+    local version, err = kong.cache:get("plugins_plan:version", TTL_ZERO, utils.uuid)
+    if err then
+      return nil, "failed to retrieve plugins plan version: " .. err
+    end
+
+    if not plugins or plugins.version ~= version then
+
+      local timeout = 60
+      if kong.configuration.database == "cassandra" then
+        -- cassandra_timeout is defined in ms
+        timeout = kong.configuration.cassandra_timeout / 1000
+
+      elseif kong.configuration.database == "postgres" then
+        -- pg_timeout is defined in ms
+        timeout = kong.configuration.pg_timeout / 1000
+      end
+      local plugins_mutex_opts = {
+        name = "plugins",
+        timeout = timeout,
+      }
+
+      local ok, err = concurrency.with_coroutine_mutex(plugins_mutex_opts, function()
+        -- we have the lock but we might not have needed it. check the
+        -- version again and rebuild if necessary
+        version, err = kong.cache:get("plugins_plan:version", TTL_ZERO, utils.uuid)
+        if err then
+          return nil, "failed to re-retrieve plugins plan version: " .. err
+        end
+
+        if not plugins or plugins.version ~= version then
+          local ok, err = build_plugins(version)
+          if not ok then
+            return nil, "error found when building plugins plan: " .. err
+          end
+        end
+
+        return true
+      end)
+      if not ok then
+        return nil, "failed to rebuild plugins plan: " .. err
+      end
+    end
+
+    return true
+  end
+
+  get_plugins = function()
+    return plugins
+  end
 end
 
 
@@ -623,11 +778,34 @@ local function balancer_setup_stage2(ctx)
 end
 
 
+local function set_init_versions_in_cache()
+  local ok, err = kong.cache:get("router:version", TTL_ZERO, function()
+    return "init"
+  end)
+  if not ok then
+    return nil, "could not set router version in cache: " .. tostring(err)
+  end
+
+  local ok, err = kong.cache:get("plugins_plan:version", TTL_ZERO, function()
+    return "init"
+  end)
+  if not ok then
+    return nil, "could not set plugins plan version in cache: " .. tostring(err)
+  end
+
+  return true
+end
+
+
 -- in the table below the `before` and `after` is to indicate when they run:
 -- before or after the plugins
 return {
   build_router = build_router,
 
+  build_plugins = build_plugins,
+  update_plugins = update_plugins,
+  get_plugins = get_plugins,
+  set_init_versions_in_cache = set_init_versions_in_cache,
   -- exported for unit-testing purposes only
   _set_check_router_rebuild = _set_check_router_rebuild,
